Include header for object size standard type
The header previously only defined ssize_t for the platform(s) which
lacked a standard definition in sys/types.h (namely Windows), but then
relied on this being available by some other inclusion on POSIX
compliant platforms (namely Linux). This caused a compilation error if
only ecl_util.h is included.

Include the header in which it is defined so it will always be available
regardless of what else is included. The guard in the header files means
that no existing code should notice any difference.

Also, define a variants for Windows which reflects the word size of the
platform that is used.