Fixed missing MSW branches
The link_count seems to be used only to detect the start segments of the
different branches. In that way the link_count must _not_ be incremented
if a segment on a different branch is reffering to the segment.
If we did, we would be missing all branches that has an endpoint
continuing in another branch.