ERT-988: Changed to use BaseCEnum
Changed the implementation of the Ecl enums EclFileEnum, EclTypeEnum,
EclPhaseEnum and EclFileFlagEnum to use BaseCEnum. Observe that a
consequency of this change is that the the old style
'ecl.ECL_FLOAT_TYPE' will not longer work.

Have also removed the ecl_type attribute of the EclKW class, asking the
C layer each time instead.