ERT-1009: Addedd class forward_load_context
Add a class forward_load_context which holds on to all the information
needed/used while loading the results from the forward model. The
absolutely most important element of the forward_load_context is the
target_fs which is used to assert size of the GEN_DATA
elements. The forward_load_context allows us to remove the enkf_fs
member from the gen_data_config object.