Simplified JobQueue.submit().
The JobQueue.submit() method used to return a job object, but that
object could not be properly initialized before the job was actually
submitted to the backend queue - which could be deferred for a long
time. Simplified by just returning integer queu_index. Removed several
broken/stale Python list classes.