Write nexus data to correct ecl timesteps
Before this patch, nexus data written to ecl was assumed to start at
timestep 0. In ecl, all keywords and variables are present from timestep
0. In nexus, however, instances of regions, well, etc. are not present
in the dataset before the timestep at which the first appear. This
caused instances that appear at a later timestep to start at timestep 0.
Now the data is written at the timestep it appears.