Test timesteps of instances not present initially
In nexus, instances aren't present in the dataset before the first timestep
they appare. In the dataset SPE1_delayedWell.plt, the well instance "3" does
not appare before timestep number 5. In ecl, however, the well is present from
the beginning. Because of this, the indices where the data starts is different
in the nex and ecl representations.