Write nexus data to correct ecl timesteps (#250)
* Write nexus data to correct ecl timesteps

Before this patch, nexus data written to ecl was assumed to start at
timestep 0. In ecl, all keywords and variables are present from timestep
0. In nexus, however, instances of regions, well, etc. are not present
in the dataset before the timestep at which the first appear. This
caused instances that appear at a later timestep to start at timestep 0.
Now the data is written at the timestep it appears.

* Rename smspec_nodes to ecl_values

* Test timesteps of instances not present initially

In nexus, instances aren't present in the dataset before the first timestep
they appare. In the dataset SPE1_delayedWell.plt, the well instance "3" does
not appare before timestep number 5. In ecl, however, the well is present from
the beginning. Because of this, the indices where the data starts is different
in the nex and ecl representations.