New ecl3 fortio implementation
A new implementation of the basic fortran-style IO implementation, based
around the four functions sizeof, skip, get, and put. These functions
report all errors in a recoverable manner, and attempts to roll-back
changes as much as possible to be robust and predictable.

The cmakedefine for the htons family of functions is replaced by
test_big_endian and a generator-expression, to be more obvious in the
makefile what actually depends on it, and that the definition is not
exported.