Add EOF+UNEXPECTED_EOF, use in get/put/skip/sizeof
Add the new error codes EOF and UNEXPECTED_EOF for more precisely
describing why the primitive get/put functions fails.

The tests require some endianness handling when creating test data,
which in this patch is a copy of the implementation in the library. This
is expected to change, once the need for endianness management is
properly understood.