Plug memory leak in set_gen_keys
The smspec_alloc functions sometimes return a heap-allocated cstring,
which is implicitly converted to a std::string when assigned to gen_key1
and gen_key2. This leads to a leak of the returned cstring as the
std::string takes a copy, and forgets about its source.

Remove this opportunity by always storing cstrings in a unique_ptr.