Merge ecl3/f77 into ecl3/keyword
The functionality in the f77 module is sufficiently trivial, and targets
largely the same level of abstraction as they ecl3/keyword module, that
it makes sense for them to merge.

In this, the elements-in-block function is replaced by the assumption
that ecl3_get_native does the job instead, and that the caller figures
out number-of-elements through division.