Set CMAKE_MACOSX_RPATH on if not defined
Policy CMP0042 was introduced in CMake 3.0 and concerns the default
behavior of the target property MACOSX_RPATH, ie OFF before 3.0 and ON
after 3.0

It is possible that, at a certain point in time, libecl (ert) required
to stick with the old behavior. It is also possible that the code and
the cmake files were changed over time, without proper testing on
MacOSX. It appears that sticking with the old behavior has become
unnecessary, so we are explicitly setting the new default value for
older versions of CMake as well.

This, in turn, allows us to install multiple versions of libecl on the
same machine (eg, in different virtual environments) without resorting
to DYLD_LIBRARY_PATH (which is kind of broken from OSX El Capitan on)
asd