Remove relics of ecl3
The goal was to gradually move to a new version, ecl3, and re-implement
ecl2 using ecl3 primitives. This work has been abandoned, in favour of a
new library [1] that does not attempt to support ecl2 features and
behaviour.

Since ecl3 is abanadoned (in this repository), the code is dead and fit
for removal.

[1] http://github.com/equinor/ecl3