Fix create_3d for coarse groups
There was an issue where create_3d would not correctly handle keywords
containing only active values for coarse groups. This seems to be
because mapping between cell indecies and active indecies are no longer
valid for the kind of assignment done in create_3d. The fix is to look
up the active index when create_3d is given a keyword containing only
values for active cells.