Check ecl_sum data is not empty
Instead of changing the function prototype and thus breaking API, we
take advantage of the existing undefined behaviour when trying to grab
the last element of an empty dataset. We throw a C++ exception and then
catch it at the appropriate location. This does not modify the API or ABI.