Rewrite test_grid_equinor_large_case
Instead of reading a large grdecl file stored internally, the large case
test now writes a moderately large grdecl file to a temporary path.

This test was introduced with commit c07d948d6fd0ffc299560048dc1c7ccd65663777
which also added handling of index overflow in reading grdecl file.
However, the grdecl file in the test is not actually large enough
to hit that overflow handling, likely because that would take
up too much space on disk and too much time in testing. Instead
it loads a grdecl file that is reasonably large (~4gb).

To speed up testing we reduce the size by a factor of ~8.