Simplify util_abort using backward-cpp and fmtlib
`util_abort` uses the `backtrace` functionality of GCC along with an
external program to create the stacktrace. This is both really slow and
non portable, only really working on Linux and macOS. This new
implementation uses the well regarded
[backward-cpp](https://github.com/bombela/backward-cpp) library to
create the stack traces. I also added the popular formatting library
[fmt](https://github.com/fmtlib/fmt), which allows us Python format
string-like formatting in C++, with type correctness and everything.

This is accomplished by using the [Conan](https://conan.io) package
manager for C++.