Throw sensible error if invalid internal_index
The report_map used in ecl_sum_data_iget_report_end when asking
for a report_range is populated with steps from the restart file. In the
scenario where there are more steps in the restart file than in the
summary file, retrieving one of the report_step will return -1.