Fix a buffer overflow in restart summary read
The number of blocks of 8 character strings in the RESTART keyword can
now be more than 9, and in any case the file could always have the
length set to something larger, leading to stack corruption.

Also fixed that it was always assumed that the restart filename
was relative.

Before, many restart cases would have been ignored because the file
was too long (or incorrectly assumed to be relative). Now, restart
data will most likely be included, which could lead to unexpected changed
behavior.