Remove broken function Layer.copy
This function has thrown an IndexError on the parameters
since 45e1c3828208e29ea8ac4bc95eaeb34157cad2e4 (7 years ago as of this
commit). We have not gotten a bug report of this. We can therefore be pretty sure nobody is using it.
In order to work properly, the call to _copy should not have the first argument layer.